Fixed import statements, renamed to word_cloud
