+++ lib/presentation/views/home/home.dart
@@ -5,8 +5,7 @@ import 'package:sis_group6/core/theme/app.dart';
 import 'package:sis_group6/mock/sentiment_data.dart';
 import 'package:sis_group6/presentation/widgets/sentiment_chart.dart';
 import 'package:sis_group6/presentation/widgets/tweet_list_card.dart';
-
-import '../../widgets/word_cloud.dart';
+import 'package:sis_group6/presentation/widgets/word_cloud.dart';
 
 class HomeView extends StatelessWidget {
   const HomeView({Key? key}) : super(key: key);
